requirement.txt
appdirs==1.4.4
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-eventtools==1.0.0
django-js-asset==1.2.2
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
draftjs-exporter==2.1.7
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
html5lib==1.0.1
icalendar==4.0.5
idna==2.9
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.8.0
Pillow==6.2.2
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
ratelim==0.1.6
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.8.1
wagtail-cache==1.0.0
wagtail-import-export==0.2
wagtailfontawesome==1.2.1
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.3
wrapt==1.11.2
